fix(runtime): drop redundant u32 cast in the Linux peer-uid path
libc::ucred::uid is already u32 on Linux, so the cast tripped
clippy::unnecessary_cast under -D warnings. The macOS branch reads
into a libc::uid_t local, which clippy does not resolve to a concrete
type, so this only ever failed on the ubuntu leg.

+++ crates/khive-runtime/src/daemon.rs
@@ -358,7 +358,7 @@ fn peer_uid(stream: &UnixStream) -> std::io::Result<u32> {
         if rc != 0 {
             return Err(std::io::Error::last_os_error());
         }
-        Ok(cred.uid as u32)
+        Ok(cred.uid)
     }
 
     #[cfg(not(any(
